+++ example/lib/main.dart
@@ -1,9 +1,8 @@
 import 'package:clickio_consent_sdk/clickio_consent_sdk.dart';
+import 'package:clickio_consent_sdk_example/widgets/primary_button.dart';
 import 'package:flutter/material.dart';
 
-void main() {
-  runApp(const MyApp());
-}
+void main() => runApp(MaterialApp(home: const MyApp()));
 
 class MyApp extends StatefulWidget {
   const MyApp({super.key});
@@ -27,99 +26,79 @@ class _MyAppState extends State<MyApp> {
 
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
-      home: Scaffold(
-        appBar: AppBar(
-          centerTitle: true,
-          title: Padding(
-            padding: const EdgeInsets.only(top: 16.0),
-            child: const Text(
-              'Clickio Consent SDK',
-              style: TextStyle(fontWeight: FontWeight.w600),
-            ),
+    final padding16 = SizedBox(height: 16);
+    final padding4 = SizedBox(height: 4);
+
+    return Scaffold(
+      appBar: AppBar(
+        centerTitle: true,
+        title: Padding(
+          padding: const EdgeInsets.only(top: 16.0),
+          child: const Text(
+            'Clickio Consent SDK',
+            style: TextStyle(fontWeight: FontWeight.w600),
           ),
         ),
-        body: SafeArea(
-          child: Column(
-            children: [
-              Center(
-                child: Column(
-                  children: [
-                    const SizedBox(height: 16),
-                    ElevatedButton(
-                      onPressed: openDialogInDefaultMode,
-                      style: ElevatedButton.styleFrom(
-                        backgroundColor: Color.fromRGBO(12, 20, 67, 1.0),
-                      ),
-                      child: const Text(
-                        'Default Mode',
-                        style: TextStyle(color: Colors.white),
-                      ),
+      ),
+      body: SafeArea(
+        child: Builder(
+          builder:
+              (context) => Column(
+                children: [
+                  Center(
+                    child: Column(
+                      children: [
+                        padding16,
+                        PrimaryButton(
+                          title: 'Default Mode',
+                          onTap: openDialogInDefaultMode,
+                        ),
+                        padding4,
+                        PrimaryButton(
+                          title: 'Resurface Mode',
+                          onTap: openDialog,
+                        ),
+                        padding4,
+                        PrimaryButton(
+                          title: 'Custom URL',
+                          onTap: webViewLoadUrl,
+                        ),
+                        padding4,
+                        PrimaryButton(
+                          title: 'Get Consent Data',
+                          onTap: getConsentData,
+                        ),
+                      ],
                     ),
-                    const SizedBox(height: 4),
-                    ElevatedButton(
-                      onPressed: openDialog,
-                      style: ElevatedButton.styleFrom(
-                        backgroundColor: Color.fromRGBO(12, 20, 67, 1.0),
-                      ),
-                      child: const Text(
-                        'Resurface Mode',
-                        style: TextStyle(color: Colors.white),
-                      ),
-                    ),
-                    const SizedBox(height: 4),
-                    ElevatedButton(
-                      onPressed: getConsentData,
-                      style: ElevatedButton.styleFrom(
-                        backgroundColor: const Color.fromRGBO(12, 20, 67, 1.0),
-                      ),
-                      child: const Text(
-                        'Get Consent Data',
-                        style: TextStyle(color: Colors.white),
-                      ),
-                    ),
-                    // const SizedBox(height: 4),
-                    // ElevatedButton(
-                    //   onPressed: () {},
-                    //   style: ElevatedButton.styleFrom(
-                    //     backgroundColor: const Color.fromRGBO(12, 20, 67, 1.0),
-                    //   ),
-                    //   child: const Text(
-                    //     'Clear cache',
-                    //     style: TextStyle(color: Colors.white),
-                    //   ),
-                    // ),
-                  ],
-                ),
-              ),
-              const SizedBox(height: 16),
-              Expanded(
-                child: SingleChildScrollView(
-                  child: Column(
-                    children:
-                        _consentData.entries.map((entry) {
-                          return ListTile(
-                            title: Text(
-                              entry.key,
-                              style: TextStyle(
-                                fontWeight: FontWeight.bold,
-                                fontSize: 18,
-                              ),
-                            ),
-                            subtitle: Text(
-                              entry.value.toString(),
-                              style: TextStyle(
-                                fontWeight: FontWeight.w500,
-                                fontSize: 16,
-                              ),
-                            ),
-                          );
-                        }).toList(),
                   ),
-                ),
+                  padding16,
+                  Expanded(
+                    child: SingleChildScrollView(
+                      child: Column(
+                        children:
+                            _consentData.entries.map((entry) {
+                              return ListTile(
+                                title: Text(
+                                  entry.key,
+                                  style: TextStyle(
+                                    fontWeight: FontWeight.bold,
+                                    fontSize: 18,
+                                  ),
+                                ),
+                                subtitle: Text(
+                                  entry.value.toString(),
+                                  style: TextStyle(
+                                    fontWeight: FontWeight.w500,
+                                    fontSize: 16,
+                                  ),
+                                ),
+                              );
+                            }).toList(),
+                      ),
+                    ),
+                  ),
+                ],
               ),
-            ],
-          ),
         ),
       ),
     );
@@ -142,16 +121,60 @@ class _MyAppState extends State<MyApp> {
     );
   }
 
+  Future<void> webViewLoadUrl() async {
+    final backgroundColor = const Color.fromARGB(255, 227, 247, 246);
+
+    clickioConsentSdk.setOnWebClose(() async {
+      await Navigator.maybePop(context);
+      await clickioConsentSdk.cleanup();
+    });
+
+    final webView = clickioConsentSdk.webViewLoadUrl(
+      url: 'https://example.com',
+      webViewConfig: WebViewConfig(
+        backgroundColor: backgroundColor,
+        height: 600,
+        width: 350,
+        gravity: WebViewGravity.center,
+      ),
+    );
+
+    await showDialog(
+      context: context,
+      builder:
+          (_) => Dialog(
+            backgroundColor: backgroundColor,
+            insetPadding: EdgeInsets.zero,
+            clipBehavior: Clip.hardEdge,
+            child: Column(
+              mainAxisSize: MainAxisSize.min,
+              crossAxisAlignment: CrossAxisAlignment.end,
+              children: [
+                // Custom close button
+                IconButton(
+                  icon: Icon(Icons.close),
+                  onPressed: () async {
+                    // Removing WebView by button click
+                    await Navigator.maybePop(context);
+                  },
+                ),
+                // Custom WebView
+                webView,
+              ],
+            ),
+          ),
+    );
+  }
+
   Future<void> getConsentData() async {
     _consentData = await _getConsentData();
 
-    setState(() {
-      _consentData = _sortConsentData(_consentData);
-    });
+    setState(() => _consentData = _sortConsentData(_consentData));
   }
 
   Future<Map<String, String?>> _getConsentData() async {
     final sdk = clickioConsentSdk;
+
     final futures = {
       'checkConsentScope': sdk.getConsentScope(),
       'checkConsentState': sdk.getConsentState(),
@@ -171,13 +194,13 @@ class _MyAppState extends State<MyApp> {
       'getConsentedNonTcfPurposes': sdk.getConsentedNonTcfPurposes(),
     };
 
-    final entries = await Future.wait(
+    final results = await Future.wait(
       futures.entries.map(
         (entry) async => MapEntry(entry.key, await entry.value),
       ),
     );
 
-    return Map.fromEntries(entries);
+    return Map.fromEntries(results);
   }
 
   Map<String, String?> _sortConsentData(Map<String, String?> consentData) {