+++ lib/enums/enums.dart
@@ -1,2 +1,3 @@
 export './dialog_mode.dart';
 export './logs_mode.dart';
+export './web_view_gravity.dart';